refactor: split 3 files >500 LOC, update WG-127 status, fix clippy warnings
- Split agentfs.rs (531→244): extract test module to agentfs_tests.rs
- Split relationships.rs (525→402): extract row helpers to relationships_helpers.rs
- Split schema.rs (507→446): extract monitoring DDL to schema_monitoring.rs
- Fix clippy manual-let-else warnings in agentfs.rs
- Update ROADMAP_ACTIVE.md, GOAP_STATE.md, GOALS.md, ACTIONS.md for WG-127 completion
