+++ plans/GOAP_STATE.md
@@ -1,6 +1,6 @@
 # GOAP State Snapshot
 
-- **Last Updated**: 2026-05-17 (PR #552 dead_code→0; cascade tier attribution fix committed)
+- **Last Updated**: 2026-05-17 (WG-127 completed; file size splits)
 - **Version**: `0.1.31` (workspace, released)
 - **Branch**: `main`
 - **Validation**: `plans/STATUS/VALIDATION_LATEST.md`
